pyproject.toml
[tool.poetry]
name = "focus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^12.0"
asyncio = "^3.4.3"
torch = "^1.10"
pyannote-audio = "^2.1"
speechbrain = "^0.5"
sentence-transformers = "^3.0.0"
faster-whisper = "^1.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
